fix community routing to chat
